Solving issues
