+++ src/components/Cards/AboutCard/AboutCard.css
@@ -1,5 +1,6 @@
 .about__card__container{
   width: 600px;
+  
 }
 
 .about__card__overlay{
