--- src/pages/Users/NormalUser.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-export default function NormalUser(props) {
-  return (
-    <div className="Container">
-      <h1>
-        AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
-      </h1>
-    </div>
-  );
-}
