--- src/pages/donate/donation.css
@@ -1,7 +1,0 @@
-.Container {
-  margin-top: 100px !important;
-}
-
-.donate__card {
-  height: 30%;
-}
