--- src/pages/donate/donation.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-import PricingContainer from "../../components/PricingSection/PricingSection";
-import "../donate/donation.css";
-
-export default function Donation(props) {
-  return (
-    <div className="Container">
-      <PricingContainer />
-    </div>
-  );
-}
