+++ src/pages/donate/index.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import PricingContainer from "../../components/PricingSection/PricingSection";
+import "../donate/style.css";
+
+export default function Donation(props) {
+  return (
+    <div className="Container">
+      <PricingContainer />
+    </div>
+  );
+}
