+++ src/pages/donate/style.css
@@ -1,0 +1,7 @@
+.Container {
+  margin-top: 100px !important;
+}
+
+.donate__card {
+  height: 30%;
+}
