--- src/pages/home/home.css
@@ -1,3 +1,0 @@
-body {
-  background-color: var(--text-color);
-}
